Extract signer subscription
