+++ packages/bridge-ui/src/main.ts
@@ -1,12 +1,14 @@
 import './app.css';
 import App from './App.svelte';
 import { Buffer } from 'buffer';
+import { setupI18n } from './i18n';
+
+globalThis.Buffer = Buffer;
+
+setupI18n({ withLocale: 'en' });
 
 const app = new App({
   target: document.getElementById('app'),
 });
-
-// @ts-ignore
-window.Buffer = Buffer;
 
 export default app;
